Support multiple measurements of the same type in BLE data
Previously if we had more than 1 HA BLE count type in an
HA BLE advertisement, the later count valve in the advertisement
would overwrite the earlier value and only a sing count
entity would be created.

This adds an _<index> to HA BLE returned data, and adds the
capability for named measurements with _<index> appendicies
to be recognised in sensor.py and each to become a separate
entity in Home Assistant.

Note: This is a breaking change for HA BLE as the _<index>
is always appended even if there is only one of a particular
measurement; entity names in Home Assistant will change.
With a little extra code in ha_ble.py we could avoid this
and only update to add "_0" when the second instance of a
measurement appears.